fix circular import by restructuring textformat module
