+++ insitupy/utils/utils.py
@@ -18,91 +18,6 @@ from insitupy._constants import (XENIUM_HEX_TO_INT_CONV_DICT,
                                  XENIUM_INT_TO_HEX_CONV_DICT)
 
 
-class textformat:
-    '''
-    Helper class to format printed text.
-    e.g. print(color.RED + color.BOLD + 'Hello, World!' + color.END)
-    '''
-    # colors and formats
-    # PURPLE = '\033[95m'
-    # CYAN = '\033[96m'
-    # DARKCYAN = '\033[36m'
-    # BLUE = '\033[94m'
-    # GREEN = '\033[92m'
-    # YELLOW = '\033[93m'
-    # RED = '\033[91m'
-    # BOLD = '\033[1m'
-    # UNDERLINE = '\033[4m'
-    # END = '\033[0m'
-
-    ResetAll = "\033[0m"
-
-    Bold       = "\033[1m"
-    Dim        = "\033[2m"
-    Underlined = "\033[4m"
-    Blink      = "\033[5m"
-    Reverse    = "\033[7m"
-    Hidden     = "\033[8m"
-
-    ResetBold       = "\033[21m"
-    ResetDim        = "\033[22m"
-    ResetUnderlined = "\033[24m"
-    ResetBlink      = "\033[25m"
-    ResetReverse    = "\033[27m"
-    ResetHidden     = "\033[28m"
-
-    Default      = "\033[39m"
-    Black        = "\033[30m"
-    Red          = "\033[31m"
-    Green        = "\033[32m"
-    Yellow       = "\033[33m"
-    Blue         = "\033[34m"
-    Magenta      = "\033[35m"
-    Cyan         = "\033[36m"
-    LightGray    = "\033[37m"
-    DarkGray     = "\033[90m"
-    LightRed     = "\033[91m"
-    LightGreen   = "\033[92m"
-    LightYellow  = "\033[93m"
-    LightBlue    = "\033[94m"
-    Purple       = "\033[95m"
-    LightCyan    = "\033[96m"
-    White        = "\033[97m"
-
-    BackgroundDefault      = "\033[49m"
-    BackgroundBlack        = "\033[40m"
-    BackgroundRed          = "\033[41m"
-    BackgroundGreen        = "\033[42m"
-    BackgroundYellow       = "\033[43m"
-    BackgroundBlue         = "\033[44m"
-    BackgroundMagenta      = "\033[45m"
-    BackgroundCyan         = "\033[46m"
-    BackgroundLightGray    = "\033[47m"
-    BackgroundDarkGray     = "\033[100m"
-    BackgroundLightRed     = "\033[101m"
-    BackgroundLightGreen   = "\033[102m"
-    BackgroundLightYellow  = "\033[103m"
-    BackgroundLightBlue    = "\033[104m"
-    BackgroundLightMagenta = "\033[105m"
-    BackgroundLightCyan    = "\033[106m"
-    BackgroundWhite        = "\033[107m"
-
-    # signs
-    TSIGN = "\u251c"
-    LSIGN = "\u2514"
-    HLINE = "\u2500"
-    RARROWHEAD = "\u27A4"
-    TICK = "\u2714"
-    CIRCLE_EMPTY = "\u25EF"
-    CIRCLE_FILLED = "\u2B24"
-    CIRCLE_HALF = "\u25D0"
-    CIRCLE_THREEQUARTER = "\u25D5"
-    CIRCLE_ONEQUARTER = "\u25D4"
-    CROSS = "\u2716"
-
-    # spacer
-    SPACER = "    "
-
 def create_ansi_color_code_from_rgb(rgb_color):
     # Create the ANSI escape code
     ansi_escape_code = f'\033[38;2;{rgb_color[0]};{rgb_color[1]};{rgb_color[2]}m'
